renamed DictionaryGenerator -> TextParserToFrequencyDictionary
