+++ TagsCloudContainer/App/CloudVisualizer/CloudVisualizer.cs
@@ -2,7 +2,7 @@
 using TagsCloudContainer.App.Settings;
 using TagsCloudContainer.Infrastructure.CloudGenerator;
 using TagsCloudContainer.Infrastructure.CloudVisualizer;
-using TagsCloudContainer.Infrastructure.DictionaryGenerator;
+using TagsCloudContainer.Infrastructure.TextParserToFrequencyDictionary;
 using TagsCloudContainer.Infrastructure.DataReader;
 
 namespace TagsCloudContainer.App.CloudVisualizer
@@ -10,15 +10,15 @@ namespace TagsCloudContainer.App.CloudVisualizer
     internal class CloudVisualizer : ICloudVisualizer
     {
         private readonly ICloudGenerator cloudGenerator;
-        private readonly IFrequencyDictionaryGenerator dictionaryGenerator;
+        private readonly ITextParserToFrequencyDictionary textParserToTextParserToFrequencyDictionary;
         private readonly IImageGenerator imageGenerator;
         private readonly IDataReader inputDataReader;
 
-        public CloudVisualizer(IDataReader inputDataReader, IFrequencyDictionaryGenerator dictionaryGenerator,
+        public CloudVisualizer(IDataReader inputDataReader, ITextParserToFrequencyDictionary textParserToTextParserToFrequencyDictionary,
             ICloudGenerator cloudGenerator, IImageGenerator imageGenerator)
         {
             this.inputDataReader = inputDataReader;
-            this.dictionaryGenerator = dictionaryGenerator;
+            this.textParserToTextParserToFrequencyDictionary = textParserToTextParserToFrequencyDictionary;
             this.cloudGenerator = cloudGenerator;
             this.imageGenerator = imageGenerator;
         }
@@ -26,8 +26,8 @@ namespace TagsCloudContainer.App.CloudVisualizer
         public Bitmap Visualize(string inputFileName, ImageSettings imageSettings)
         {
             var lines = inputDataReader.ReadLines();
-            var dictionary = dictionaryGenerator.GenerateFrequencyDictionary(lines);
-            var cloud = cloudGenerator.GenerateCloud(dictionary);
+            var frequencyDictionary = textParserToTextParserToFrequencyDictionary.GenerateFrequencyDictionary(lines);
+            var cloud = cloudGenerator.GenerateCloud(frequencyDictionary);
             return imageGenerator.GenerateImage(cloud);
         }
     }
