+++ TagsCloudContainer/App/TextParserToFrequencyDictionary/EnumerableExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using TagsCloudContainer.Infrastructure.DictionaryGenerator;
+using TagsCloudContainer.Infrastructure.TextParserToFrequencyDictionary;
 
-namespace TagsCloudContainer.App.FrequencyDictionaryGenerator
+namespace TagsCloudContainer.App.TextParserToFrequencyDictionary
 {
     internal static class EnumerableExtensions
     {
