+++ TagsCloudContainer/App/TextParserToFrequencyDictionary/SimpleWordFilter.cs
@@ -1,6 +1,6 @@
-﻿using TagsCloudContainer.Infrastructure.DictionaryGenerator;
+﻿using TagsCloudContainer.Infrastructure.TextParserToFrequencyDictionary;
 
-namespace TagsCloudContainer.App.FrequencyDictionaryGenerator
+namespace TagsCloudContainer.App.TextParserToFrequencyDictionary
 {
     internal class SimpleWordFilter : IWordFilter
     {
