+++ TagsCloudContainer/Infrastructure/TextParserToFrequencyDictionary/IWordFilter.cs
@@ -1,4 +1,4 @@
-﻿namespace TagsCloudContainer.Infrastructure.DictionaryGenerator
+﻿namespace TagsCloudContainer.Infrastructure.TextParserToFrequencyDictionary
 {
     internal interface IWordFilter
     {
